pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cpred"
version = "0.1.0"
description = "Prediction of viable circular permutation cleavage sites in proteins"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.82",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "torch>=2.0",
    "networkx>=3.1",
    "pandas>=2.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov"]

[project.scripts]
cpred = "cpred.cli:main"

[tool.setuptools.packages.find]
include = ["cpred*"]
